refactor: 统一网站元数据获取接口
